Move roundtrip test into separate file; check collection too
