+++ rslib/src/import_export/package/colpkg/export.rs
@@ -317,58 +317,3 @@ impl MediaFileWriter {
         Ok(self)
     }
 }
-
-#[cfg(test)]
-mod test {
-    use tempfile::tempdir;
-
-    use super::*;
-    use crate::{import_export::package::import_colpkg, media::MediaManager};
-
-    fn collection_with_media(dir: &Path, name: &str) -> Result<Collection> {
-        let name = format!("{name}_src");
-        let media_folder = dir.join(format!("{name}.media"));
-        std::fs::create_dir(&media_folder)?;
-        let col = CollectionBuilder::new(dir.join(format!("{name}.anki2")))
-            .set_media_paths(media_folder, dir.join(format!("{name}.mdb")))
-            .build()?;
-        let mgr = MediaManager::new(&col.media_folder, &col.media_db)?;
-        let mut ctx = mgr.dbctx();
-        mgr.add_file(&mut ctx, "1", b"1")?;
-        mgr.add_file(&mut ctx, "2", b"2")?;
-        mgr.add_file(&mut ctx, "3", b"3")?;
-        Ok(col)
-    }
-
-    #[test]
-    fn roundtrip() -> Result<()> {
-        let _dir = tempdir()?;
-        let dir = _dir.path();
-
-        for (legacy, name) in [(true, "legacy"), (false, "v3")] {
-            // export to a file
-            let col = collection_with_media(dir, name)?;
-            let colpkg_name = dir.join(format!("{name}.colpkg"));
-            col.export_colpkg(&colpkg_name, true, legacy, |_| ())?;
-            // import into a new collection
-            let anki2_name = dir
-                .join(format!("{name}.anki2"))
-                .to_string_lossy()
-                .into_owned();
-            let import_media_dir = dir.join(format!("{name}.media"));
-            import_colpkg(
-                &colpkg_name.to_string_lossy(),
-                &anki2_name,
-                import_media_dir.to_str().unwrap(),
-                |_| Ok(()),
-            )?;
-
-            // confirm media imported correctly
-            assert_eq!(std::fs::read(import_media_dir.join("1"))?, b"1");
-            assert_eq!(std::fs::read(import_media_dir.join("2"))?, b"2");
-            assert_eq!(std::fs::read(import_media_dir.join("3"))?, b"3");
-        }
-
-        Ok(())
-    }
-}
